alpha/requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Django==2.0.1
django-widget-tweaks==1.4.1
gunicorn==19.7.1
idna==2.6
Pillow==5.0.0
pkg-resources==0.0.0
psycopg2==2.7.4
python-decouple==3.1
pytz==2018.3
requests==2.18.4
urllib3==1.22
